services/rule-based-service/pyproject.toml
[project]
name = "rule-based-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "stanza>=1.11.0",
    "ufal-morphodita>=1.11.2.1",
    "uralicnlp>=2.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
